+++ src/main/java/com/smartpotatoo/map_your_trip_be/entity/schedule/SchedulesDateRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SchedulesDateRepository extends JpaRepository<SchedulesDateEntity, Long> {
+public interface SchedulesDateRepository extends JpaRepository<SchedulesDateEntity, Integer> {
     // Optional<SchedulesDateEntity> findByScheduleIdAndDate(Long schedulesId, String date);
-    List<SchedulesDateEntity> findByScheduleId(Long scheduleId);
+    List<SchedulesDateEntity> findByScheduleId(int scheduleId);
 }
