+++ src/main/java/com/smartpotatoo/map_your_trip_be/entity/schedule/SchedulesRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SchedulesRepository extends JpaRepository<SchedulesEntity, Long> {
+public interface SchedulesRepository extends JpaRepository<SchedulesEntity, Integer> {
 
 }
